feat: goal mixin updated
